# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tire Management System API",
        "description": "API documentation for the Tire Management System",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/inspections": {
            "get": {
                "tags": [
                    "Inspection"
                ],
                "summary": "List all inspections",
                "operationId": "cdb1fd832563e68e25989020d115f480",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of inspections"
                    }
                }
            },
            "post": {
                "tags": [
                    "Inspection"
                ],
                "summary": "Create inspection task",
                "operationId": "4540ab73a00c3c035bc66624ea7395a1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "vehicle_id",
                                    "type"
                                ],
                                "properties": {
                                    "vehicle_id": {
                                        "type": "integer"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "scheduled_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "assigned_to": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inspection task created"
                    }
                }
            }
        },
        "/api/inspections/{id}": {
            "get": {
                "tags": [
                    "Inspection"
                ],
                "summary": "Get inspection details",
                "operationId": "93c79566c217ffe8b5c2e701a1ee8be2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Inspection details"
                    }
                }
            },
            "put": {
                "tags": [
                    "Inspection"
                ],
                "summary": "Edit inspection task",
                "operationId": "5054c55501e3ee65f66285289614144b",
                "responses": {
                    "200": {
                        "description": "Updated"
                    }
                }
            }
        },
        "/api/inspections/{id}/assign": {
            "post": {
                "tags": [
                    "Inspection"
                ],
                "summary": "Assign inspection to inspector",
                "operationId": "87d5b373cf59e1e8b882c1be172ec4ab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Assigned"
                    }
                }
            }
        },
        "/api/inspections/{id}/schedule": {
            "post": {
                "tags": [
                    "Inspection"
                ],
                "summary": "Schedule inspection",
                "operationId": "e3dafc9beded7656fc39383d0480e42b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "scheduled_at"
                                ],
                                "properties": {
                                    "scheduled_at": {
                                        "type": "string",
                                        "format": "date-time"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Scheduled"
                    }
                }
            }
        },
        "/api/inspections/{id}/review": {
            "post": {
                "tags": [
                    "Inspection"
                ],
                "summary": "Submit inspection for review",
                "operationId": "212e5503689e54ebc8f773e0fbd12714",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Under Review"
                    }
                }
            }
        },
        "/api/inspections/{id}/approve": {
            "post": {
                "tags": [
                    "Inspection"
                ],
                "summary": "Approve inspection",
                "operationId": "e3ded5d3a89bec6905b4982f1b3a2644",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Approved"
                    }
                }
            }
        },
        "/api/inspections/{id}/reject": {
            "post": {
                "tags": [
                    "Inspection"
                ],
                "summary": "Reject inspection",
                "operationId": "9dc706440481c2e3a9641497f194ce19",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reason"
                                ],
                                "properties": {
                                    "reason": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rejected"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}